feat: add support for zip files (#27)
* feat: support for .zip files

* feat: zip viewer

* feat: full screen view

* feat: use react-images for the lightbox

+++ src/utils/fileUtils.ts
@@ -1,4 +1,5 @@
-import { unlink, readdirSync } from 'fs';
+/* eslint-disable import/prefer-default-export */
+import { unlink, readdirSync, lstatSync, rmdirSync } from 'fs';
 
 const path = require('path');
 
@@ -8,10 +9,15 @@ export const purgeDecryptedFiles = (dir: string) => {
     filenames.map((filename) => {
       const fullPath = path.join(dir, filename);
       return new Promise((resolve, reject) => {
-        unlink(fullPath, (error) => {
-          if (error) reject(error);
+        if (lstatSync(fullPath).isDirectory()) {
+          rmdirSync(fullPath, { recursive: true });
           resolve(filename);
-        });
+        } else {
+          unlink(fullPath, (error) => {
+            if (error) reject(error);
+            resolve(filename);
+          });
+        }
       });
     })
   );
